Crate wit_component

The WebAssembly component tooling.

Modules

metadata

Definition for encoding of custom sections within core wasm modules of component-model related data.

Structs

ComponentEncoder

An encoder of components based on wit interface definitions.

WitPrinter

A utility for printing WebAssembly interface definitions to a string.

Enums

DecodedWasm

Result of the decode function.

StringEncoding

Supported string encoding formats.

Functions

decode

Decodes an in-memory WebAssembly binary into a WIT Resolve and associated metadata.

encode

Encodes the given package within resolve to a binary WebAssembly representation.